Create a subclass for checking plugin requirements
