+++ algolia-woo-indexer.php
@@ -16,9 +16,17 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+/**
+ * Main class file
+ */
 require_once plugin_dir_path( __FILE__ ) . '/classes/class-algolia-woo-indexer.php';
 
-$algowooindexer = \ALGOWOO\Algolia_Woo_Indexer::get_instance();
+/**
+ * Class for checking plugin requirements
+ */
+require_once plugin_dir_path( __FILE__ ) . '/classes/class-check-requirements.php';
+
+$algowooindexer = \Algowoo\Algolia_Woo_Indexer::get_instance();
 
 register_activation_hook( __FILE__, array( $algowooindexer, 'activate_plugin' ) );
 register_deactivation_hook( __FILE__, array( $algowooindexer, 'deactivate_plugin' ) );
